+++ Demo/Demo/ContentView.swift
@@ -17,176 +17,204 @@ struct ContentView: View {
     @State var isDelayedPresented = false
     @State var isPopoverPresented = false
     @State var isSheetPresented = false
-    @State var isPageViewPresented = false
-    @State var isSlideshowPresented = false
     @State var isToolbarPopoverPresented = false
     
     @State var pageViewPageIndex = 0
     @State var slideshowPageIndex = 0
-    
+
     var body: some View {
         NavigationStack {
             List {
                 Section("Onboardings") {
-                    listButton(
-                        .popover, 
-                        presentPopover,
-                        "Onboarding popover",
-                        "This will only be presented once"
-                    )
-                    .demoPopover($isPopoverPresented)
-                    
-                    listButton(
-                        .sheet,
-                        presentSheet,
-                        "Onboarding modal",
-                        "This will only be presented once"
-                    )
-                    .sheet(isPresented: $isSheetPresented) {
-                        Text("Use modal onboarding screens to show more complex onboardings, like page views or slideshows."
-                        )
-                        .padding()
-                        .multilineTextAlignment(.center)
-                    }
-                    
-                    VStack(spacing: 20) {
-                        listButton(
-                            .conditional,
-                            presentConditional,
-                            "Conditional onboarding",
-                            "This will only be presented once, and only if the toggle below is enabled."
-                        )
-                        Toggle(isOn: $isConditionalEnabled) {}
-                            .labelsHidden()
-                            .padding(.leading, 5)
-                    }
-                    .demoPopover(
-                        $isConditionalPresented,
-                        text: "Conditional onboardings are great if a condition must be met before the onboarding is presented."
-                    )
-
-                    listButton(
-                        .delayed,
-                        presentDelayed,
-                        "Delayed onboarding",
-                        "This will only be presented once, and only after \(Onboarding.demoDelayed.requiredPresentationAttempts) taps."
-                    )
-                    .demoPopover(
-                        $isDelayedPresented,
-                        text: "Delayed onboardings are great to control the order of a list of onboardings."
-                    )
-                    
-                    listButton(
-                        .correctBehavior,
-                        presentCorrectBehavior,
-                        "Correct behavior onboarding",
-                        "This will be presented every 3rd tap."
-                    )
-                    .demoPopover(
-                        $isCorrectBehaviorPresented,
-                        text: "Correct behavior onboardings are great to teach users how to behave, e.g. in games."
-                    )
+                    popoverButton
+                    sheetButton
+                    conditionalButton
+                    delayedButton
+                    behaviorButton
                 }
 
                 Section("Flows") {
-                    listButton(
-                        .popover,
-                        presentLocalizedPageView,
-                        "Present an onboarding page view",
-                        "This onboarding flow lets you swipe through a set of pages."
-                    )
-                    .sheet(isPresented: $isPageViewPresented) {
-                        DemoPageView(
-                            onboarding: .demo,
-                            index: $pageViewPageIndex
-                        )
-                    }
-
-                    listButton(
-                        .popover,
-                        presentLocalizedSlideshow,
-                        "Present an onboarding slideshow",
-                        "This onboarding flow automatically slides through a set of pages."
-                    )
-                    .sheet(isPresented: $isSlideshowPresented) {
-                        DemoSlideshow(
-                            onboarding: .demo,
-                            index: $pageViewPageIndex
-                        )
-                    }
+                    pageViewLink
+                    slideshowLink
                 }
             }
             .navigationTitle("OnboardingKit")
             .onAppear(perform: presentToolbarPopover)
             .safeAreaInset(edge: .bottom) {
-                VStack {
+                VStack(spacing: 0) {
                     Divider()
-                    Button("Reset onboarding state", action: reset)
-                        .buttonStyle(.borderedProminent)
-                        .padding()
+                    OnboardingPrimaryButton(
+                        "Reset onboarding state",
+                        action: reset
+                    )
+                    .padding()
                 }
-                .frame(maxWidth: .infinity)
                 .background(.thinMaterial)
             }
             .toolbar {
                 ToolbarItem(placement: .confirmationAction) {
                     toolbarButton
-                        .demoPopover($isToolbarPopoverPresented)
                 }
             }
         }
     }
 }
 
-private extension View {
+// MARK: - Buttons
 
-    func listButton(
-        _ icon: Image,
-        _ action: @escaping () -> Void,
-        _ title: String,
-        _ text: String
-    ) -> some View {
-        Button(action: action) {
-            Label(
-                title: {
-                    LabeledContent {} label: {
-                        Text(title)
-                        Text(text)
-                    }
-                },
-                icon: {
-                    icon
-                }
-            )
-            .contentShape(Rectangle())
-        }
-        .buttonStyle(.plain)
-    }
-
-    func demoPopover(
-        _ isPresented: Binding<Bool>,
-        text: String = "Use popovers hints to inform the user about important parts of the UI."
-    ) -> some View {
-        self.popover(isPresented: isPresented) {
-            Text(text)
-                .padding()
-                .padding(.vertical)
-                #if os(iOS)
-                .fixedSize(horizontal: false, vertical: true)
-                .presentationCompactAdaptation(.none)
-                #endif
-        }
-    }
-}
-
-@MainActor
 private extension ContentView {
-    
+
+    var behaviorButton: some View {
+        Button(action: presentCorrectBehavior) {
+            listLabel(
+                .correctBehavior,
+                "Correct behavior popover",
+                "This will be presented every 3rd tap."
+            )
+        }
+        .onboardingPopover(
+            "Correct behavior onboardings are great to teach users how to behave, e.g. in games.",
+            isPresented: $isCorrectBehaviorPresented
+        )
+    }
+
+    var conditionalButton: some View {
+        Button(action: presentConditional) {
+            HStack(spacing: 20) {
+                listLabel(
+                    .conditional,
+                    "Conditional onboarding",
+                    "This will only be presented once, if the toggle is enabled."
+                )
+                Toggle(isOn: $isConditionalEnabled) {}
+                    .labelsHidden()
+            }
+        }
+        .onboardingPopover(
+            "Conditional onboardings are great if a condition must be met before the onboarding is presented.",
+            isPresented: $isConditionalPresented
+        )
+    }
+
+    var delayedButton: some View {
+        Button(action: presentDelayed) {
+            listLabel(
+                .delayed,
+                "Delayed popover",
+                "This will only be presented once, and only after \(Onboarding.demoDelayed.requiredPresentationAttempts) taps."
+            )
+        }
+        .onboardingPopover(
+            "Use popovers hints to inform the user about important parts of the UI.",
+            isPresented: $isDelayedPresented
+        )
+    }
+
+    var pageViewLink: some View {
+        NavigationLink {
+            DemoOnboardingScreen(type: .pageView)
+        } label: {
+            listLabel(
+                .pageView,
+                "Page View",
+                "Step through a number of pages."
+            )
+        }
+    }
+
+    var popoverButton: some View {
+        Button(action: presentPopover) {
+            listLabel(
+                .popover,
+                "Onboarding popover",
+                "This will only be presented once"
+            )
+        }
+        .onboardingPopover(
+            "Use popovers hints to inform the user about important parts of the UI.",
+            isPresented: $isPopoverPresented
+        )
+    }
+
+    var sheetButton: some View {
+        Button(action: presentSheet) {
+            listLabel(
+                .sheet,
+                "Onboarding sheet",
+                "This will only be presented once"
+            )
+        }
+        .sheet(isPresented: $isSheetPresented) { sheet }
+    }
+
+    var sheet: some View {
+        Text("Use modal onboarding screens to show more complex onboardings, like page views or slideshows."
+        )
+        .padding()
+        .multilineTextAlignment(.center)
+    }
+
+    var slideshowLink: some View {
+        NavigationLink {
+            DemoOnboardingScreen(type: .slideshow)
+        } label: {
+            listLabel(
+                .slideshow,
+                "Slideshow",
+                "Slide through a number of pages."
+            )
+        }
+    }
+
     var toolbarButton: some View {
         Button(action: presentToolbarPopover) {
             Image.popover.label("Show popover")
         }
+        .onboardingPopover(
+            "Use popovers hints to inform the user about important parts of the UI.",
+            isPresented: $isToolbarPopoverPresented
+        )
     }
+}
+
+// MARK: - Views
+
+private extension View {
+
+    func listButton<Content: View>(
+        _ action: @escaping () -> Void,
+        _ content: @escaping () -> Content
+    ) -> some View {
+        Button(action: action) {
+            content()
+        }
+        .buttonStyle(.plain)
+    }
+
+    func listLabel(
+        _ icon: Image,
+        _ title: String,
+        _ text: String
+    ) -> some View {
+        Label(
+            title: {
+                LabeledContent {} label: {
+                    Text(title)
+                    Text(text)
+                }
+            },
+            icon: {
+                icon
+            }
+        )
+        .contentShape(Rectangle())
+    }
+}
+
+// MARK: - Functions
+
+@MainActor
+private extension ContentView {
 
     func presentConditional() {
         Onboarding.demoConditional { isConditionalEnabled }
@@ -203,18 +231,9 @@ private extension ContentView {
             .tryPresent { isDelayedPresented = true }
     }
 
-    func presentLocalizedPageView() {
-        isPageViewPresented = true
-    }
-
-    func presentLocalizedSlideshow() {
-        isSlideshowPresented = true
-    }
-
     func presentPopover() {
-        Onboarding.demoPopover.tryPresent {
-            isPopoverPresented = true
-        }
+        Onboarding.demoPopover
+            .tryPresent { isPopoverPresented = true }
     }
     
     func presentSheet() {